saving performance history

+++ historical_performance/models.py
@@ -47,7 +47,7 @@ class PerformancePortfolio(models.Model):
 
 	date = models.DateField(verbose_name="Date")
 	allocation = models.ForeignKey(Allocation, verbose_name="Portfolio", related_name='performances', on_delete=models.PROTECT)
-	unit_value = models.IntegerField(verbose_name="Value")
+	unit_value = models.DecimalField(verbose_name="Value", max_digits=8, decimal_places=2)
 
 	def __str__(self):
 		return '{} - Date: {}'.format(self.portfolio.allocation, self.date.strftime("%d/%m/%Y"))
